refactor: table
