+++ components/RequestsTable/RequestsTable.tsx
@@ -6,38 +6,24 @@ import {
   RequestStatus as RequestStatusType,
   Tenant,
 } from '@/lib/db/schema';
-import {
-  useReactTable,
-  getCoreRowModel,
-  flexRender,
-  Row,
-  Cell,
-  getSortedRowModel,
-  getPaginationRowModel,
-} from '@tanstack/react-table';
+import { Row, Cell } from '@tanstack/react-table';
 import { useAuth } from '@/hooks/useAuth';
-import {
-  DateCell,
-  ResolveCell,
-  RequestTypeCell,
-  DeclineReasonCell,
-  TenantCell,
-} from './cells/Cell';
-import SaveOfferCell from './cells/SaveOfferCell';
+import { DateCell, RequestTypeCell, TenantCell } from './cells/Cell';
+import DeclineReasonCell from '@/components/RequestsTable/cells/DeclineReasonCell';
+import ResolveCell from '@/components/RequestsTable/cells/ResolveCell';
+import SaveOfferCell from '@/components/RequestsTable/cells/SaveOfferCell';
 import ReportButton from './ReportButton';
 import RequestRow from './Row';
 import { generateCustomerInfoColumns } from './table.utils';
-import clsx from 'clsx';
 import RequestStatus from '../RequestStatus/RequestStatus';
 import EmptyRequestsState from './EmptyTable';
 import { FaCheckCircle } from 'react-icons/fa';
 import { FaCircleXmark } from 'react-icons/fa6';
 import { getTenants } from '@/lib/api/tenant';
 import { useQuery } from '@tanstack/react-query';
-import { CustomColumnMeta } from '@/constants/app.types';
 import CTACell from './cells/CTACell';
-import { TablePagination } from '@/components/ui/pagination';
 import RequestDrawer from '../RequestDetails/RequestDrawer';
+import DataTable from '../ui/table';
 
 interface Props {
   requests: Request[];
@@ -140,13 +126,14 @@ const RequestsTable: FC<Props> = ({
       cell: ({
         getValue,
         cell,
+        row,
       }: {
         getValue: () => string;
         cell: Cell<Request, boolean>;
         row: Row<Request>;
       }) => {
         if (isProviderUser) {
-          return <ResolveCell cell={cell} />;
+          return <ResolveCell cell={cell} row={row} />;
         }
         const value = getValue();
         if (value === null) {
@@ -181,7 +168,11 @@ const RequestsTable: FC<Props> = ({
             tenant => tenant.id === row.original.providerTenantId,
           );
           return (
-            <DeclineReasonCell cell={cell} provider={provider as Tenant} />
+            <DeclineReasonCell
+              cell={cell}
+              provider={provider as Tenant}
+              row={row}
+            />
           );
         }
 
@@ -202,26 +193,6 @@ const RequestsTable: FC<Props> = ({
       : []),
   ];
 
-  const paginationSettings = {
-    pageIndex: 0,
-    pageSize: 10,
-  };
-  const [pagination, setPagination] = useState(paginationSettings);
-  const shouldRenderPagination = requests.length > paginationSettings.pageSize;
-
-  const table = useReactTable({
-    data: requests,
-    columns,
-    getCoreRowModel: getCoreRowModel(),
-    getSortedRowModel: getSortedRowModel(),
-    getPaginationRowModel: getPaginationRowModel(),
-    state: {
-      sorting: defaultSort,
-      pagination,
-    },
-    onPaginationChange: setPagination,
-  });
-
   if (!userData) return null;
 
   if (requests.length === 0) {
@@ -229,64 +200,23 @@ const RequestsTable: FC<Props> = ({
   }
 
   return (
-    <div className="grid grid-cols-1 gap-4 overflow-hidden">
-      <div className="overflow-x-auto">
-        <table className="min-w-full divide-y divide-gray-200">
-          <thead className="border-b border-gray-200">
-            {table.getHeaderGroups().map(headerGroup => (
-              <tr key={headerGroup.id}>
-                {headerGroup.headers.map(header => {
-                  const meta = header.column.columnDef.meta as CustomColumnMeta;
-                  const isHighlightable = meta?.isHighlightable;
-                  const width = header.column.getSize();
-                  const headerClassName = clsx(
-                    'p-4 whitespace-nowrap',
-                    {
-                      'bg-yellow-50': isHighlightable,
-                    },
-                    meta?.className ?? 'text-left',
-                  );
-                  return (
-                    <th
-                      key={header.id}
-                      className={headerClassName}
-                      style={{ minWidth: `${width}px` }}
-                    >
-                      {header.isPlaceholder
-                        ? null
-                        : flexRender(
-                            header.column.columnDef.header,
-                            header.getContext(),
-                          )}
-                    </th>
-                  );
-                })}
-              </tr>
-            ))}
-          </thead>
-          <tbody>
-            {table.getRowModel().rows.map(row => (
-              <RequestRow key={row.id} row={row} toggleDrawer={toggleDrawer} />
-            ))}
-          </tbody>
-        </table>
-      </div>
-
-      <div className="w-full mt-4">
-        {shouldRenderPagination && (
-          <TablePagination
-            currentPage={pagination.pageIndex + 1}
-            totalPages={table.getPageCount()}
-            onPageChange={page => table.setPageIndex(page - 1)}
-          />
-        )}
+    <>
+      <div className="h-full">
+        <DataTable
+          data={requests}
+          columns={columns}
+          defaultSort={defaultSort}
+          EmptyComponent={EmptyComponent}
+          onRowClick={toggleDrawer}
+          RowComponent={RequestRow}
+        />
       </div>
       <RequestDrawer
         isOpen={isDrawerOpen}
         onClose={() => setIsDrawerOpen(false)}
         request={selectedRequest}
       />
-    </div>
+    </>
   );
 };
