--- constants/app.types.ts
@@ -1,5 +1,0 @@
-export interface CustomColumnMeta {
-  isHighlightable?: boolean;
-  isCustomerInfo?: boolean;
-  className?: string;
-}
